Move calculations module to assyst; rework high level run functions a bit
